requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
bidict==0.22.1
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.5
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-engineio==4.8.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-socketio==5.10.0
python-telegram-bot==20.6
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
wsproto==1.2.0
